reformat code and mplus feature done

+++ app/src/main/java/com/BlizzardArmory/model/warcraft/mythicplusleaderboards/leaderboards/index/CurrentLeaderboards.kt
@@ -3,9 +3,9 @@ package com.BlizzardArmory.model.warcraft.mythicplusleaderboards.leaderboards.in
 import com.BlizzardArmory.model.common.Key
 import com.google.gson.annotations.SerializedName
 
-data class CurrentLeaderboards (
+data class CurrentLeaderboards(
 
-	@SerializedName("key") val key : Key,
-	@SerializedName("name") val name : String,
-	@SerializedName("id") val id : Int
+    @SerializedName("key") val key: Key,
+    @SerializedName("name") val name: String,
+    @SerializedName("id") val id: Long
 )
